requirements.txt
# Core dependencies for Narrative Consistency Auditor
# Install with: pip install -r requirements.txt
peft==0.17.1
# Text processing
tiktoken==0.8.0

# Embeddings and semantic search - stable compatible versions
torch>=2.0.0
transformers==4.49.0
sentence-transformers==2.2.2
faiss-cpu==1.8.0
# For GPU support, replace faiss-cpu with: faiss-gpu==1.8.0

# LLM integration
google-generativeai==0.8.3

# Machine Learning
scikit-learn==1.3.2
numpy==1.26.4
pandas==2.0.3

# Environment management
python-dotenv==1.0.1

# Optional: Pathway for data orchestration
# pathway==0.15.4

# Progress bars and utilities
tqdm==4.67.1
